feat: enhance user management with validation and access control

+++ src/controllers/userController.ts
@@ -2,19 +2,6 @@ import status from 'http-status';
 import { NextFunction, Request, Response } from 'express';
 import * as userService from '../services/userService';
 
-export const getAllUsers = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const users = await userService.getAllUsers();
-    res.status(status.OK).json(users);
-  } catch (err) {
-    next(err);
-  }
-};
-
 export const getUserById = async (
   req: Request,
   res: Response,
@@ -34,6 +21,10 @@ export const updateUser = async (
   next: NextFunction,
 ) => {
   try {
+    if (req.user.id !== req.params.id) {
+      return res.status(403).json({ message: 'Acesso negado' });
+    }
+
     const user = await userService.updateUser(req.params.id, req.body);
     res.status(status.OK).json(user);
   } catch (err) {
@@ -47,6 +38,10 @@ export const deleteUser = async (
   next: NextFunction,
 ) => {
   try {
+    if (req.user.id !== req.params.id) {
+      return res.status(403).json({ message: 'Acesso negado' });
+    }
+
     await userService.deleteUser(req.params.id);
     res.status(status.ACCEPTED).json({ message: 'Usuário deletado' });
   } catch (err) {
